requirements.txt
anyio == 4.2.0       
azure-core == 1.29.6      
azure-storage-blob == 12.19.0     
beautifulsoup4 == 4.12.2      
blinker == 1.7.0       
certifi == 2023.11.17  
cffi == 1.16.0
charset-normalizer == 3.3.2
click == 8.1.7
colorama == 0.4.6
cryptography == 41.0.7
Flask == 3.0.0
idna == 3.6
isodate == 0.6.1
itsdangerous == 2.1.2
Jinja2 == 3.1.2
joblib == 1.3.2
MarkupSafe == 2.1.3
mysql-connector-python == 8.2.0
numpy == 1.26.2
pandas == 2.1.4
pip == 23.3.2
protobuf == 4.21.12
pycparser == 2.21
python-dateutil == 2.8.2
pytz == 2023.3.post1
requests == 2.31.0
scikit-learn == 1.3.2
scipy == 1.11.4
setuptools == 69.0.2
six == 1.16.0
sniffio == 1.3.0
soupsieve == 2.5
threadpoolctl == 3.2.0
typing_extensions == 4.9.0
tzdata == 2023.3
urllib3 == 2.1.0
Werkzeug == 3.0.1
wheel == 0.42.0
azure-cognitiveservices-speech==1.34.0
azure-storage-blob==12.19.0
backports.entry-points-selectable==1.1.0
distlib==0.3.2
filelock==3.0.12
gunicorn==20.1.0
nltk==3.6.2
platformdirs==2.2.0
regex==2021.4.4
textblob==0.15.3
tqdm==4.61.1
virtualenv==20.7.0